\docType{package}
\name{urbstatdata-package}
\alias{urbstatdata}
\alias{urbstatdata-package}
\title{urbstatdata: Seven Urban and Built-Environment Data Sets}
\description{
Provides seven fixed, documented data sets. Each data-set help page records
the unit of observation, variables, source, licence, transformations and
limitations. The package exports no functions.
}
\keyword{datasets}
